♻️ refactor: group private modules under packages
Mirror the intent grouping of the C tree on the Python side. The public
extract, query and clean modules become packages that own their private
helpers, and the cross-cutting helpers move to a new _internal package.

Public import paths are unchanged: turbohtml.extract et al. stay
importable via each package __init__, and every name re-exported from
turbohtml stays put. Only underscore-private module paths move, with
every in-repo reference updated.
