+++ src/turbohtml/__init__.py
@@ -4,12 +4,6 @@ from __future__ import annotations
 
 from importlib.metadata import version
 
-from . import (
-    _cssmin,  # noqa: F401  # registers the CSSMinify record type with the C core on import
-    _jsminify,  # noqa: F401  # registers the JSMinify record type with the C core on import
-)
-from ._article import Article  # registers the Article record type with the C core on import
-from ._feed import Entry, Feed  # registers the Feed/Entry record types with the C core on import
 from ._html import (
     Axis,
     CData,
@@ -43,19 +37,25 @@ from ._html import (
     tokenize,
     unescape,
 )
-from ._links import Link  # registers the Link record type with the C core on import
-from ._locations import SourceLocation, SourceSpan  # registers the source-location record types on import
-from ._render import Canonical, Html, Markdown, PlainText
-from ._selectors import SelectorSyntaxError  # registers the selector error type with the C core on import
-from ._structured_data import (  # registers the JSON-LD parser and record classes with the C core on import
+from ._internal import (
+    _cssmin,  # noqa: F401  # registers the CSSMinify record type with the C core on import
+    _jsminify,  # noqa: F401  # registers the JSMinify record type with the C core on import
+)
+from ._internal._locations import SourceLocation, SourceSpan  # registers the source-location record types on import
+from ._internal._render import Canonical, Html, Markdown, PlainText
+from ._internal._selectors import SelectorSyntaxError  # registers the selector error type with the C core on import
+from .build import E, ElementMaker
+from .extract._article import Article  # registers the Article record type with the C core on import
+from .extract._feed import Entry, Feed  # registers the Feed/Entry record types with the C core on import
+from .extract._links import Link  # registers the Link record type with the C core on import
+from .extract._structured_data import (  # registers the JSON-LD parser and record classes with the C core on import
     MicrodataItem,
     OpenGraph,
     RdfaItem,
     StructuredData,
 )
-from ._xpath import XPathString  # registers the smart-string type with the C core on import
-from .build import E, ElementMaker
 from .mutations import MutationObserver, MutationRecord  # registers the MutationRecord type on import
+from .query._xpath import XPathString  # registers the smart-string type with the C core on import
 from .traverse import NodeFilter, NodeIterator, TreeWalker
 
 __version__ = version("turbohtml")
