+++ src/turbohtml/_internal/__init__.py
@@ -1,0 +1,1 @@
+"""Cross-cutting helpers shared by more than one turbohtml subsystem."""
