+++ src/turbohtml/_stubs/dom.pyi
@@ -4,12 +4,12 @@ from enum import Enum, IntEnum
 from re import Pattern
 from typing import TypeAlias, final
 
-from turbohtml._article import Article as Article
-from turbohtml._feed import Feed
-from turbohtml._links import Link
-from turbohtml._locations import SourceLocation
-from turbohtml._render import Canonical, Html, Markdown, PlainText
-from turbohtml._structured_data import JSONValue, MicrodataItem, OpenGraph, RdfaItem, StructuredData
+from turbohtml._internal._locations import SourceLocation
+from turbohtml._internal._render import Canonical, Html, Markdown, PlainText
+from turbohtml.extract._article import Article as Article
+from turbohtml.extract._feed import Feed
+from turbohtml.extract._links import Link
+from turbohtml.extract._structured_data import JSONValue, MicrodataItem, OpenGraph, RdfaItem, StructuredData
 from turbohtml.mutations import MutationRecord
 
 _Filter: TypeAlias = str | Pattern[str] | bool | Callable[[str | None], bool] | list[_Filter]
