+++ tools/fuzz/_targets.py
@@ -28,7 +28,7 @@ import turbohtml
 from turbohtml import clean
 
 # the harness targets the private URL/IDNA C bindings (th_url_split/join/percent, th_url_to_ascii) by design
-from turbohtml._urls import (  # noqa: PLC2701
+from turbohtml.extract._urls import (  # noqa: PLC2701
     _url_join,
     _url_percent_decode,
     _url_percent_encode,
